feat(exec): windows process runner

+++ src/main/java/com/aws/greengrass/util/platforms/windows/WindowsExec.java
@@ -7,6 +7,8 @@ package com.aws.greengrass.util.platforms.windows;
 
 import com.aws.greengrass.util.Exec;
 import com.aws.greengrass.util.Utils;
+import com.aws.greengrass.util.platforms.Platform;
+import com.aws.greengrass.util.platforms.UserPlatform;
 import org.zeroturnaround.process.Processes;
 
 import java.io.File;
@@ -21,8 +23,8 @@ import javax.annotation.Nullable;
 
 @SuppressWarnings("PMD.AvoidCatchingThrowable")
 public class WindowsExec extends Exec {
-    public static final String PATHEXT_KEY = "PATHEXT";
-
+    private static final String PATHEXT_KEY = "PATHEXT";
+    private static final String LOCAL_DOMAIN = ".";
     private static final List<String> PATHEXT;  // ordered file extensions to try, when no extension is provided
 
     static {
@@ -67,25 +69,33 @@ public class WindowsExec extends Exec {
 
     @Override
     public String[] getCommand() {
-        String[] decorated = cmds;
+        String[] decorated = Arrays.copyOf(cmds, cmds.length);
+        for (int i = 0; i < decorated.length; i++) {
+            final String arg = decorated[i];
+            // Space and \t require quoting otherwise will be split to more than one arg
+            if (!isQuoted(arg) && arg.matches(".*[ \t].*")) {
+                decorated[i] = "\"" + arg + "\"";
+            }
+        }
         if (shellDecorator != null) {
             decorated = shellDecorator.decorate(decorated);
-        }
-        // First item in the command is the executable. If it's given as absolute path, add quotes around it
-        // in case the path contains space which will break the whole command line
-        // See security remarks:
-        // https://docs.microsoft.com/en-us/windows/win32/api/winbase/nf-winbase-createprocesswithlogonw
-        if (isAbsolutePath(decorated[0])) {
-            decorated[0] = String.format("\"%s\"", decorated[0]);
         }
         return decorated;
     }
 
     @Override
     protected Process createProcess() throws IOException {
-        ProcessBuilder pb = new ProcessBuilder();
-        pb.environment().putAll(environment);
-        return pb.directory(dir).command(getCommand()).start();
+        if (needToSwitchUser()) {
+            WindowsRunasProcess winProcess = new WindowsRunasProcess(LOCAL_DOMAIN, userDecorator.getUser());
+            winProcess.setAdditionalEnv(environment);
+            winProcess.setCurrentDirectory(dir.getAbsolutePath());
+            winProcess.start(String.join(" ", getCommand()));
+            return winProcess;
+        } else {
+            ProcessBuilder pb = new ProcessBuilder();
+            pb.environment().putAll(environment);
+            return pb.directory(dir).command(getCommand()).start();
+        }
     }
 
     @Override
@@ -96,12 +106,16 @@ public class WindowsExec extends Exec {
         if (process == null || !process.isAlive()) {
             return;
         }
-        Process killerProcess = new ProcessBuilder().command("taskkill", "/f", "/t", "/pid",
-                Integer.toString(Processes.newPidProcess(process).getPid())).start();
+        int pidToKill = process instanceof WindowsRunasProcess ? ((WindowsRunasProcess) process).getPid()
+                : Processes.newPidProcess(process).getPid();
+        Process killerProcess =
+                new ProcessBuilder().command("taskkill", "/f", "/t", "/pid", Integer.toString(pidToKill)).start();
         try {
-            killerProcess.waitFor();
-            process.destroyForcibly();
-            process.waitFor(5, TimeUnit.SECONDS);
+            int taskkillExitCode = killerProcess.waitFor();
+            if (taskkillExitCode != 0) {
+                process.destroyForcibly();
+                process.waitFor(5, TimeUnit.SECONDS);
+            }
         } catch (InterruptedException e) {
             Thread.currentThread().interrupt();
         } finally {
@@ -109,7 +123,24 @@ public class WindowsExec extends Exec {
         }
     }
 
+    /**
+     * Returns true if we need to create process as another user. Otherwise, just use ProcessBuilder.
+     */
+    private boolean needToSwitchUser() throws IOException {
+        if (userDecorator == null) {
+            return false;
+        }
+        // check if same as current user
+        UserPlatform.UserAttributes currUser = Platform.getInstance().lookupCurrentUser();
+        return !(currUser.getPrincipalName().equals(userDecorator.getUser()) || currUser.getPrincipalIdentifier()
+                .equals(userDecorator.getUser()));
+    }
+
     private static boolean isAbsolutePath(String p) {
         return new File(p).isAbsolute();
+    }
+
+    private static boolean isQuoted(String s) {
+        return s.startsWith("\"") && s.endsWith("\"") && !s.endsWith("\\\"");
     }
 }
